+++ config/nova-page-manager.php
@@ -4,13 +4,27 @@ return [
 
   /*
   |--------------------------------------------------------------------------
-  | Table Name
+  | Table name
   |--------------------------------------------------------------------------
   |
   | Set a custom table for Nova Page Manager to store its data.
   |
   */
 
-  'table' => 'nova_page_manager'
+  'table' => 'nova_page_manager',
+
+
+  /*
+  |--------------------------------------------------------------------------
+  | Max locales shown on index
+  |--------------------------------------------------------------------------
+  |
+  | Sets the number of locales shown on index. If the number of locales
+  | exceeds the defined count, the locales will be shown only on the detail
+  | view.
+  |
+  */
+
+  'max_locales_shown_on_index' => 2,
 
 ];