+++ src/Nova/Page.php
@@ -28,8 +28,9 @@ class Page extends TemplateResource
         $tableName = NovaPageManager::getPagesTableName();
         $templateClass = $this->getTemplateClass();
         $templateFields = $this->getTemplateFields();
-        $localeParentField = LocaleParentField::make("Translations");
-        if (count(NovaPageManager::getLocales()) > 2) {
+        $localeParentField = LocaleParentField::make('Translations');
+
+        if (count(NovaPageManager::getLocales()) > config('nova-page-manager.max_locales_shown_on_index', 2)) {
             $localeParentField = $localeParentField->hideFromIndex();
         }