+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.0]
+
+### Added
+
+- Added config option `max_locales_shown_on_index` with default value of `2`
+
+### Changed
+
+- Translation options are now shown on both Index and Detail views
+- When the amount of translation options exceeds the configured limit, the options are only displayed on the Detail view (thanks to [@slovenianGooner](https://github.com/slovenianGooner))
+
+## [1.5.2]
+
+### Changed
+
+- Fixed SEO field imports in `Page` class from [@slovenianGooner](https://github.com/slovenianGooner)
+
 ## [1.5.1]
 
 ### Changed
@@ -111,6 +128,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Programmatically created templates for pages and regions
 - Multilanguage support
 
+[1.6.0]: https://github.com/optimistdigital/nova-page-manager/compare/1.5.2...1.6.0
+[1.5.2]: https://github.com/optimistdigital/nova-page-manager/compare/1.5.1...1.5.2
 [1.5.1]: https://github.com/optimistdigital/nova-page-manager/compare/1.5.0...1.5.1
 [1.5.0]: https://github.com/optimistdigital/nova-page-manager/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/optimistdigital/nova-page-manager/compare/1.3.1...1.4.0